requirements.txt
fastapi
openai
uvicorn
python-dotenv
packaging
requests
